CAUSEWAY-3794: separate out table hinting concerns
